test(react): share the stub main-thread environment between the two first-frame tests
Both the root `<Background>` and the root `<MainThread>` runtime tests
evaluate a real `main-thread.js` in a VM against stubbed element PAPI; the
stubs are the same and belong in one place.
